update selections order
