+++ app/components/constraints_component.html.erb
@@ -1,0 +1,30 @@
+<%= content_tag @tag || :div, id: @id, class: @classes do %>
+  <% if @render_headers %>
+    <h2 class="sr-only visually-hidden"><%= t('blacklight.search.search_constraints_header') %></h2>
+  <% end %>
+
+  <% if @render_headers %>
+    <span class="constraints-label d-flex align-self-center fw-bold"><%= t('blacklight.search.filters.title') %></span>
+  <% end %>
+  <% if query_constraints_area.present? %>
+    <% query_constraints_area.each do |constraint| %>
+      <%= constraint %>
+    <% end %>
+  <% else %>
+    <%= query_constraints %>
+  <% end %>
+
+  <% if facet_constraints_area.present? %>
+    <% facet_constraints_area.each do |constraint| %>
+      <%= constraint %>
+    <% end %>
+  <% else %>
+    <%= facet_constraints %>
+  <% end %>
+
+  <% additional_constraints.each do |constraints| %>
+    <%= constraints %>
+  <% end %>
+
+  <%= render @start_over_component.new if @start_over_component %>
+<% end %>
