+++ spec/components/start_over_button_component_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe Blacklight::StartOverButtonComponent, type: :component do
+RSpec.describe StartOverButtonComponent, type: :component do
   subject(:render) { render_inline(component) }
 
   let(:component) { described_class.new }
@@ -11,6 +11,6 @@ RSpec.describe Blacklight::StartOverButtonComponent, type: :component do
     render
     # We override core Blacklight components to always set the search_field to 'all_fields'
     expect(component.call).to include('href="/catalog?search_field=all_fields"')
-    expect(component.call).to include('Start Over')
+    expect(component.call).to include('Clear all')
   end
 end
